[depthMap] Rename hpp/cpp files RefineRc to Refine
